# internal/controller/openapi.client.json
{
  "components": {
    "parameters": {
      "projectRef": {
        "description": "项目 UUID、当前 slug，或未过期的 slug 别名。过期别名返回 404 PROJECT_NOT_FOUND。",
        "in": "path",
        "name": "project_ref",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
...
      "Error": {
        "description": "统一错误信封：{ error: { code, message, details } }。",
        "properties": {
          "error": {
            "description": "错误对象。",
            "properties": {
              "code": {
                "enum": [
                  "PROJECT_NOT_FOUND",
                  "VERSION_NOT_FOUND",
                  "VERSION_LINE_NOT_FOUND",
                  "VERSION_NOT_VISIBLE",
                  "VERSION_REVOKED",
                  "NO_SAFE_TARGET",
                  "INTERMEDIATE_UNAVAILABLE",
                  "MIN_OS_NOT_MET",
                  "CHANNEL_CONFLICT",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "PRECONDITION_FAILED",
                  "HW_REV_INCOMPATIBLE",
                  "RATE_LIMITED",
                  "CHANGELOG_QUERY_INVALID",
                  "NOT_FOUND",
                  "NOT_READY",
                  "INTERNAL_ERROR",
                  "INVALID_REQUEST",
                  "INVALID_QUERY_PARAM",
                  "ENGINE_MISMATCH"
                ],
                "type": "string",
                "description": "本平面会返回的错误码。code 是英文稳定值。"
              },
              "details": {
                "description": "可选细节。没有时可省略。"
              },
              "message": {
                "type": "string",
                "description": "给人看的说明，英文。不要依赖它做逻辑，请看 code。"
              }
            },
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "ProjectPublic": {
        "additionalProperties": false,
        "description": "可以给客户端看的项目设置，不含密钥。",
        "properties": {
          "compare_engine": {
            "description": "版本比较方式：semver 或 integer。",
            "type": "string"
          },
          "created_at": {
            "description": "项目创建时间（UTC）。",
            "format": "date-time",
            "type": "string"
          },
          "default_locale": {
            "description": "默认语言。",
            "type": "string"
          },
          "device_id_policy": {
            "description": "设备标识存储策略：hashed / raw / none。",
            "enum": [
              "hashed",
              "raw",
              "none"
            ],
            "type": "string"
          },
          "force_https": {
            "description": "是否强制 HTTPS。",
            "type": "boolean"
          },
          "minimum_supported_version": {
            "description": "最低支持版本；低于它会强制更新。",
            "nullable": true,
            "type": "string"
          },
          "require_client_token": {
            "description": "检查更新是否必须带项目 Token。",
            "type": "boolean"
          },
          "slug": {
            "description": "项目短名，也可用在 URL 里。",
            "type": "string"
          },
          "storage_visibility": {
            "description": "public 可直链；private 下载地址带短时签名。",
            "type": "string"
          },
          "updated_at": {
            "description": "最近修改时间（UTC）。",
            "format": "date-time",
            "type": "string"
          },
          "uuid": {
            "description": "项目 UUID。",
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "opaque",
        "description": "项目开启 require_client_token 时使用。Authorization: Bearer <项目客户端 Token>。与 X-Project-Token 是同一凭证，不是管理员登录。",
        "scheme": "bearer",
        "type": "http"
      },
      "channelTokenAuth": {
        "description": "可选渠道口令。按 SHA-256 比对，不会记入日志。填错或漏填时，check 会跳过受保护渠道，不会 403。",
        "in": "header",
        "name": "X-Channel-Token",
        "type": "apiKey"
      },
      "storeTokenAuth": {
        "description": "独立的商店协议 Token（若配置了）。用 Authorization: Bearer 或 X-Project-Token 传递。不读取查询参数。",
        "in": "header",
        "name": "X-Project-Token",
        "type": "apiKey"
      },
      "projectTokenAuth": {
        "description": "项目客户端 Token，放在 X-Project-Token 或 Authorization: Bearer。",
        "in": "header",
        "name": "X-Project-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "KiriVers 客户端平面：给软件客户端检查更新、下载安装包、校验文件、上报遥测。默认流程是 POST /update/check，有更新就下载 package_url 指向的全量或增量包。单文件差量再 POST /update/diff；多文件按文件比较增量则 GET integrity 对照本地后 POST /update/pack（同一地址轮询）。更新日志、公告、渠道/语言目录、设备 report 都是可选接口。商店协议文档走 /store/... ，不是官方 JSON 客户端。管理后台在管理平面。",
    "title": "KiriVers Client API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/projects/{project_ref}": {
      "get": {
        "description": "不含密钥。用于确认项目存在、是否强制 HTTPS、是否要带客户端 Token。开启 require_client_token 时需要 Bearer 或 X-Project-Token。",
        "parameters": [
          {
            "$ref": "#/components/parameters/projectRef"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPublic"
                }
              }
            },
            "description": "成功。"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "项目要求客户端 Token 时未提供或无效（UNAUTHORIZED）。"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "项目强制 HTTPS 但当前不是 HTTPS（FORBIDDEN）。"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "项目不存在（PROJECT_NOT_FOUND）。"
          }
        },
        "summary": "读取项目公开设置",
        "tags": [
          "Project"
        ]
      },
      "options": {
        "description": "不要求客户端 Token。Origin 匹配 cors_origins 时回写 Access-Control-Allow-*。",
        "parameters": [
          {
            "$ref": "#/components/parameters/projectRef"
          }
        ],
        "responses": {
          "204": {
            "description": "无正文。"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "项目不存在（PROJECT_NOT_FOUND）。"
          }
        },
        "summary": "CORS 预检",
        "tags": [
          "Project"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "本平面 HTTP 监听地址（客户端默认 :8080）。",
      "url": "/"
    }
  ],
...
}